requirements.txt
Flask==3.0.3
pandas==2.2.3
numpy==1.26.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.1
bs4==0.0.1
retry==0.9.2
openpyxl==3.1.2
lxml==5.1.0